requirements.txt
astroid==2.4.2
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
certifi==2020.6.20
chardet==3.0.4
idna==2.10
isort==5.6.4
jmespath==0.10.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
s3transfer==0.3.7
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.11
wrapt==1.12.1
